% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{add_noise_to_point}
\alias{add_noise_to_point}
\title{Add Noise to a Point}
\usage{
add_noise_to_point(vec, noise_var = 0.05)
}
\arguments{
\item{vec}{A numeric vector.}

\item{noise_var}{Standard deviation of noise.}
}
\value{
A noisy vector.
}
\description{
Adds independent Gaussian noise to each coordinate of a vector.
}
\examples{
add_noise_to_point(c(1, 2, 3), noise_var = 0.1)
}
