% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibril_assay.R
\name{assay}
\alias{assay}
\title{Classify Fibril Structures}
\usage{
assay(net)
}
\arguments{
\item{net}{A network object.}
}
\value{
A data frame where each row is a node and each column indicates membership in a fibril topology.
}
\description{
Takes a network and classifies each node into various fibril topology types based on graphlet orbit counts.
}
\examples{
set.seed(0)
net <- network::network.initialize(24, directed = FALSE)
net <- default_ergm_simulation(c(100, -50), "edges+kstar(2)",
             list(burn_time = 2^20, node_count = 24))
assay(net)
}
