% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{breed_generation}
\alias{breed_generation}
\title{Breed a New Generation}
\usage{
breed_generation(
  parents,
  num_lin_kids,
  num_mutants,
  noise_var = 0.1,
  use_line_density = TRUE,
  min_line_density = 0.5,
  pair_max = 10000,
  child_max = 1e+05
)
}
\arguments{
\item{parents}{List of parent vectors.}

\item{num_lin_kids}{Number of children between each parent pair.}

\item{num_mutants}{Number of mutants per child.}

\item{noise_var}{Variance of the Gaussian noise for mutation.}

\item{use_line_density}{Logical; adapt number of points to distance.}

\item{min_line_density}{Minimum line density.}

\item{pair_max}{Maximum number of parent pairs.}

\item{child_max}{Maximum number of total offspring.}
}
\value{
A list of offspring vectors.
}
\description{
Breeds a new generation by interpolating between all parent pairs and mutating children and parents.
}
\examples{
parents <- list(c(0, 0), c(1, 1), c(2, 2))
new_gen <- breed_generation(parents, num_lin_kids = 5, num_mutants = 2)
length(new_gen)

}
