% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{breed_two_parents}
\alias{breed_two_parents}
\title{Breed Two Parents}
\usage{
breed_two_parents(
  vec1,
  vec2,
  num_points_init,
  use_line_density = TRUE,
  min_line_density = 0.5
)
}
\arguments{
\item{vec1}{First parent vector.}

\item{vec2}{Second parent vector.}

\item{num_points_init}{Initial number of children points to generate.}

\item{use_line_density}{Logical; adjust number of points based on distance.}

\item{min_line_density}{Minimum allowed point density along the line.}
}
\value{
A list of offspring vectors.
}
\description{
Creates evenly spaced interpolated points between two parent vectors based on the distance between them.
}
\examples{
parent1 <- c(0, 0)
parent2 <- c(1, 1)
children <- breed_two_parents(parent1, parent2, 5)
print(children)
}
