% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{breeding}
\alias{breeding}
\title{Breeding Functions for Parameter Evolution}
\description{
Functions to create offspring parameter vectors by interpolation and mutation, supporting the genetic algorithm.
}
