% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{default_ergm_simulation}
\alias{default_ergm_simulation}
\title{Default ERGM Simulation Function}
\usage{
default_ergm_simulation(theta, suff_stats, sim_args)
}
\arguments{
\item{theta}{Parameter vector (coefficients).}

\item{suff_stats}{ERGM statistics formula string.}

\item{sim_args}{List of extra simulation arguments (burn time, etc.)}
}
\value{
A simulated network.
}
\description{
A basic ERGM simulation using ergm model
}
