% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{evolution}
\alias{evolution}
\title{Evolution Functions for Genetic Algorithm}
\description{
Functions to simulate, evaluate, and evolve parameter sets across multiple generations to optimize fibril yield in networks.
}
