% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibril_assay.R
\name{fibril_assay}
\alias{fibril_assay}
\title{Fibril Assay Functions}
\description{
Classify nodes of networks into fibril topologies using ORCA orbit counts.
}
\details{
This function calculates the yield of different fibril topologies.
It counts only nodes that are part of the interior of a fibrillar section of a graph,
leading to slight undercounting at fibril ends but encouraging longer fibrils.

For background on ORCA orbit counting, see:
\url{https://www.jstatsoft.org/article/view/v071i10/0}.
Also see ergm.graphlets package:
\url{http://CRAN.R-project.org/package=ergm.graphlets}.
}
