% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperspheroid_sampling.R
\name{get_hypersphere}
\alias{get_hypersphere}
\title{Generate Random Points in a Hypersphere}
\usage{
get_hypersphere(d, n, radius, center)
}
\arguments{
\item{d}{Number of dimensions.}

\item{n}{Number of points.}

\item{radius}{Radius of the hypersphere.}

\item{center}{Center coordinates.}
}
\value{
A matrix where each column is a sampled point.
}
\description{
Create random points inside a hypersphere of given radius and center.
}
\examples{
get_hypersphere(3, 10, 1, c(0, 0, 0))
}
