% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperspheroid_sampling.R
\name{normalize}
\alias{normalize}
\title{Normalize a Vector}
\usage{
normalize(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
Normalized vector.
}
\description{
Normalize a numeric vector to have unit L2 norm.
}
\examples{
normalize(c(3, 4))
}
