% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_evolution_script.R
\name{run_full_evolution}
\alias{run_full_evolution}
\title{Run the Full Evolution Pipeline}
\usage{
run_full_evolution(
  init_param,
  spheroid_pt_count = 50,
  radius = 2,
  node_count = 48,
  suff_stats = "edges+kstar(2)+nsp(1:2)",
  fib_type = "2-ribbon",
  num_lin_kids = 3,
  num_mutants_init = 2,
  num_initial_survivors = 8,
  num_generations = 3,
  scale_fac = 0.15,
  top_frac = 0.25,
  noise_var_init = 0.1,
  reps = 1,
  burn_time = 2^20,
  spheroid_sim_burn_time = 2^21 - 1,
  final_sim_burn_time = 2^21 - 1,
  np = 16,
  max_survivors = 4,
  fix_edge = TRUE,
  fix_edge_value = init_param[1],
  init_yield_cut = 0,
  smart_var = TRUE,
  smart_pts = TRUE,
  use_line_density = TRUE,
  min_line_density = 3.3,
  print_params = FALSE,
  pair_max = 1000,
  child_max = 20,
  final_reps = 32,
  simulation_function = default_ergm_simulation,
  sim_args = list(burn_time = burn_time, node_count = node_count),
  spheroid_sim_args = NULL,
  final_sim_args = NULL
)
}
\arguments{
\item{init_param}{Initial parameter vector to sample around.}

\item{spheroid_pt_count}{Number of initial points to sample.}

\item{radius}{Radius of the initial hyperspheroid.}

\item{node_count}{Number of nodes in networks.}

\item{suff_stats}{ERGM sufficient statistics formula (string).}

\item{fib_type}{Fibril topology to optimize.}

\item{num_lin_kids}{(default 3)}

\item{num_mutants_init}{(default 2)}

\item{num_initial_survivors}{how many values to have in gen_zero (default 8)}

\item{num_generations}{Number of generations to evolve.}

\item{scale_fac}{(default 0.15)}

\item{top_frac}{(default 0.25)}

\item{noise_var_init}{(default 0.10)}

\item{reps}{(default 16)}

\item{burn_time}{MCMC burn-in time for simulations during genetic algorithm}

\item{spheroid_sim_burn_time}{burn-in time for initial spheroid simulations}

\item{final_sim_burn_time}{burn-in time for reps of final simulation}

\item{np}{Number of parallel cores.}

\item{max_survivors}{(default 4)}

\item{fix_edge}{(default TRUE)}

\item{fix_edge_value}{(default init_param[1])}

\item{init_yield_cut}{(default 0.0)}

\item{smart_var}{(default TRUE)}

\item{smart_pts}{(default TRUE)}

\item{use_line_density}{(default TRUE)}

\item{min_line_density}{(default 3.3)}

\item{print_params}{(default FALSE)}

\item{pair_max}{(default 1000)}

\item{child_max}{(default 20)}

\item{final_reps}{(default 32)}

\item{simulation_function}{Function for simulation (default is ERGM simulation)}

\item{sim_args}{A list with extra arguments for the simulation}

\item{spheroid_sim_args}{Arguments for initial spheroid simulationns}

\item{final_sim_args}{Arguments for final simulations}
}
\value{
A list containing evolved generations and final simulation results.
}
\description{
Run the Full Evolution Pipeline
}
\examples{
\dontrun{
run_full_evolution(init_param = c(100, -30, 10, 25),
                   spheroid_pt_count = 10,
                   node_count = 48,
                   suff_stats = "edges+kstar(2)+nsp(1:2)",
                   generations = 2)
}
}
\seealso{
\code{\link{evolve}}: Description of parameters
}
