% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_log_backend}
\alias{set_log_backend}
\title{Set Logging Backend}
\usage{
set_log_backend(backend = c("base", "parallel"))
}
\arguments{
\item{backend}{One of "base" or "parallel"}
}
\description{
Set logging backend for fibrilEvolution
}
