% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_parallel}
\alias{simulate_parallel}
\title{Simulate All Parameter Sets}
\usage{
simulate_parallel(
  params,
  suff_stats,
  node_count,
  burn_time = 2^20,
  np = 16,
  simulation_function = default_ergm_simulation,
  sim_args = list(burn_time = burn_time, node_count = node_count)
)
}
\arguments{
\item{params}{List of parameter vectors.}

\item{suff_stats}{ERGM sufficient statistics formula.}

\item{node_count}{Number of nodes.}

\item{burn_time}{MCMC burn-in time.}

\item{np}{Number of parallel cores.}

\item{simulation_function}{(default default_ergm_simulation)}

\item{sim_args}{Arguments for simulation}
}
\value{
List of simulation results.
}
\description{
Simulates multiple parameter sets in parallel.
}
\examples{
simulate_parallel(list(c(1, -1), c(2, -2)), "edges+kstar(2)", 48, np = 2)
}
