% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_parallel_with_yield}
\alias{simulate_parallel_with_yield}
\title{Simulate Networks from a List of Parameters}
\usage{
simulate_parallel_with_yield(
  param_list,
  suff_stats,
  fib_type,
  node_count,
  np = 16,
  burn_time = sum(2^(1:20)),
  simulation_function = default_ergm_simulation,
  sim_args = list(burn_time = burn_time, node_count = node_count)
)
}
\arguments{
\item{param_list}{List of parameter vectors.}

\item{suff_stats}{ERGM sufficient statistics formula as a string.}

\item{fib_type}{Type of fibril topology to measure (e.g., "2-ribbon").}

\item{node_count}{Number of nodes in the network.}

\item{np}{Number of parallel cores.}

\item{burn_time}{MCMC burn-in time.}

\item{simulation_function}{Function for simulation (default default_ergm_simulation)}

\item{sim_args}{Arguments for simulation_function}
}
\value{
A list containing parameters, the statistics formula, and their corresponding fibril yields.
}
\description{
Given a list of parameter sets, simulate networks and evaluate fibril yields based on the chosen topology.
}
\examples{
simulate_parallel_with_yield(list(c(100, -30), c(101, -25)),
    "edges+kstar(2)", "2-ribbon", 48, np = 2)
}
